steps towards masking some subimages
